Move usePatronData to the global user helper file
It is a handy helper that is needed globally.
Also changed it to return all the properties from react-query in case we
need them in the future.
