+++ src/components/material/material-buttons/online/MaterialButtonOnlineInfomediaArticle.tsx
@@ -6,10 +6,9 @@ import { Manifestation } from "../../../../core/utils/types/entities";
 import { useUrls } from "../../../../core/utils/url";
 import { Button } from "../../../Buttons/Button";
 import { infomediaModalId } from "../../infomedia/InfomediaModal";
-import { isResident } from "../../../../core/utils/helpers/user";
+import { isResident, usePatronData } from "../../../../core/utils/helpers/user";
 import MaterialButtonLoading from "../generic/MaterialButtonLoading";
 import MaterialButtonDisabled from "../generic/MaterialButtonDisabled";
-import { usePatronData } from "../../helper";
 
 export interface MaterialButtonOnlineInfomediaArticleProps {
   size?: ButtonSize;
