+++ src/core/utils/helpers/user.ts
@@ -1,6 +1,7 @@
 import { isEmpty } from "lodash";
 import { Patron } from "../types/entities";
 import { hasToken } from "../../token";
+import { useGetPatronInformationByPatronIdV2 } from "../../fbs/fbs";
 
 export const isAnonymous = () => {
   return !hasToken("user");
@@ -13,5 +14,10 @@ export const isBlocked = (patron: Patron) => {
 export const isResident = (patron: Patron) => {
   return patron.resident;
 };
+
+export const usePatronData = () =>
+  useGetPatronInformationByPatronIdV2({
+    enabled: !isAnonymous()
+  });
 
 export default {};
